Moved gridftp stuff for Globus5 into globus5 folder
