--- repository/gridftp/authz/AuthzWorker.h
@@ -1,7 +1,0 @@
-#ifndef AUTHZWORKER_H
-#define AUTHZWORKER_H
-
-const char *    getVersion();
-int             checkAuthorization( char * client_id, char * object, char * action );
-
-#endif
