+++ repository/gridftp/globus5/authz/AuthzWorker.h
@@ -1,0 +1,7 @@
+#ifndef AUTHZWORKER_H
+#define AUTHZWORKER_H
+
+const char *    getVersion();
+int             checkAuthorization( char * client_id, char * object, char * action );
+
+#endif
