+++ repository/gridftp/globus5/authz/gsi-authz.conf
@@ -1,0 +1,8 @@
+GLOBUS_GSI_AUTHZ_SYSTEM_INIT           /opt/datafed/libDataFedAuthz  gsi_authz_init
+GLOBUS_GSI_AUTHZ_SYSTEM_DESTROY        /opt/datafed/libDataFedAuthz  gsi_authz_destroy
+GLOBUS_GSI_AUTHZ_HANDLE_INIT           /opt/datafed/libDataFedAuthz  gsi_authz_handle_init
+GLOBUS_GSI_AUTHORIZE_ASYNC             /opt/datafed/libDataFedAuthz  gsi_authz_authorize_async
+GLOBUS_GSI_AUTHZ_CANCEL                /opt/datafed/libDataFedAuthz  gsi_authz_cancel
+GLOBUS_GSI_AUTHZ_HANDLE_DESTROY        /opt/datafed/libDataFedAuthz  gsi_authz_handle_destroy
+GLOBUS_GSI_GET_AUTHORIZATION_IDENTITY  /opt/datafed/libDataFedAuthz  gsi_authz_identify
+globus_mapping                         /opt/datafed/libDataFedAuthz  gsi_map_user
